% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPC_colSums.R
\name{RPC_colSums}
\alias{RPC_colSums}
\title{Return counts for each (binned) value for a variable.}
\usage{
RPC_colSums(df, expl_vars)
}
\description{
Chi2 analysis is run for each variable independently. This function expects
the dataset to be 1-hot encoded. As such, we can sum over rows/columns to get
counts.
}
\details{
Return: counts for each variable (vector)
}
